+++ src/core_modules/capture-core/components/Pages/common/EnrollmentOverviewDomain/EnrollmentPageLayout/EnrollmentReadOnlyBadge/EnrollmentReadOnlyBadge.component.tsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import { ReadOnlyBadge } from '../../../../../ReadOnlyBadge';
+import { useEnrollmentAccessContext } from '../../EnrollmentAccessContext';
+
+export const EnrollmentReadOnlyBadge = () => {
+    const {
+        isEventPage,
+        currentStageWriteAccess,
+        programWriteAccess,
+        trackedEntityTypeWriteAccess,
+        anyStageWriteAccess,
+        anyStageReadAccess,
+        trackedEntityTypeName,
+        isEventWithinValidPeriod,
+        canEditCompletedEvent,
+    } = useEnrollmentAccessContext();
+
+    if (isEventPage) {
+        return (
+            <ReadOnlyBadge
+                programStageWriteAccess={currentStageWriteAccess}
+                eventWithinValidPeriod={isEventWithinValidPeriod}
+                canEditCompletedEvent={canEditCompletedEvent}
+                trackedEntityName={trackedEntityTypeName}
+                inlineLabel
+            />
+        );
+    }
+
+    const stagesEffectivelyReadOnly = !anyStageWriteAccess && anyStageReadAccess;
+    const showAllMissing = !programWriteAccess && !trackedEntityTypeWriteAccess && stagesEffectivelyReadOnly;
+    if (!showAllMissing) return null;
+
+    return (
+        <ReadOnlyBadge
+            programWriteAccess={false}
+            trackedEntityTypeWriteAccess={false}
+            programStageWriteAccess={false}
+            trackedEntityName={trackedEntityTypeName}
+            inlineLabel
+        />
+    );
+};