+++ src/core_modules/capture-core/components/WidgetEventNote/WidgetEventNote.epics.ts
@@ -43,7 +43,7 @@ export const addNoteForEventEpic = (
             const payload = action.payload;
             const eventId = state.dataEntries[payload.dataEntryId].eventId;
             const useNewEndpoint = featureAvailable(FEATURES.newNoteEndpoint);
-            const { firstName, surname, username } = CurrentUser.get();
+            const { firstName, surname } = CurrentUser.get();
             const clientId = uuid();
 
             const serverData = createServerData(eventId, payload.note, useNewEndpoint);
@@ -55,7 +55,6 @@ export const addNoteForEventEpic = (
                     surname,
                     uid: clientId,
                 },
-                storedBy: username,
                 storedAt: fromClientDate(moment().toISOString()).getServerZonedISOString(),
                 clientId,
             };