+++ src/core_modules/capture-core/components/WidgetsRelationship/WidgetTrackedEntityRelationship/NewTrackedEntityRelationship/NewTrackedEntityRelationship.container.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useState } from 'react';
-import { Button, spacers } from '@dhis2/ui';
+import { Button, spacersNum } from '@dhis2/ui';
 import { withStyles, type WithStyles } from 'capture-core-utils/styles';
 import i18n from '@dhis2/d2-i18n';
 import { NewTrackedEntityRelationshipPortal } from './NewTrackedEntityRelationship.portal';
@@ -7,11 +7,11 @@ import type { ContainerProps } from './NewTrackedEntityRelationship.types';
 
 const styles = {
     container: {
-        padding: `0 ${spacers.dp12} ${spacers.dp12} ${spacers.dp12}`,
+        padding: `${spacersNum.dp8}px ${spacersNum.dp12}px ${spacersNum.dp12}px`,
     },
 };
 
-export const NewTrackedEntityRelationshipPlain = ({
+const NewTrackedEntityRelationshipPlain = ({
     renderElement,
     teiId,
     orgUnitId,
@@ -24,6 +24,8 @@ export const NewTrackedEntityRelationshipPlain = ({
     renderTrackedEntitySearch,
     renderTrackedEntityRegistration,
     onSelectFindMode,
+    readOnly,
+    hideButton,
     classes,
 }: ContainerProps & WithStyles<typeof styles>) => {
     const [addWizardVisible, setAddWizardVisible] = useState(false);
@@ -40,13 +42,15 @@ export const NewTrackedEntityRelationshipPlain = ({
 
     return (
         <div className={classes.container}>
-            <Button
-                onClick={openAddWizard}
-                small
-                secondary
-            >
-                {i18n.t('New Relationship')}
-            </Button>
+            {!hideButton && !readOnly && (
+                <Button
+                    onClick={openAddWizard}
+                    small
+                    secondary
+                >
+                    {i18n.t('New Relationship')}
+                </Button>
+            )}
 
             {
                 addWizardVisible && (