+++ src/core_modules/capture-core/components/WidgetsRelationship/WidgetTrackedEntityRelationship/NewTrackedEntityRelationship/NewTrackedEntityRelationship.types.ts
@@ -33,6 +33,8 @@ export type ContainerProps = Readonly<{
     onCloseAddRelationship?: () => void;
     onOpenAddRelationship?: () => void;
     onSelectFindMode?: OnSelectFindMode;
+    readOnly?: boolean;
+    hideButton?: boolean;
 }>;