+++ src/core_modules/capture-core/components/WidgetsRelationship/WidgetTrackedEntityRelationship/WidgetTrackedEntityRelationship.types.ts
@@ -44,4 +44,8 @@ export type WidgetTrackedEntityRelationshipProps = {
         programId: string,
         onLinkToTrackedEntityFromSearch: OnLinkToTrackedEntityFromSearch,
     ) => React.ReactElement<any>;
+    readOnly?: boolean;
+    hideButton?: boolean;
+    accessReadOnly?: boolean;
+    hideReadOnlyBadge?: boolean;
 };